[create] subjectEntity to control what is returned to the frontend

+++ src/modules/subject/subject.service.ts
@@ -4,6 +4,7 @@ import { UpdateSubjectDto } from './dto/update-subject.dto';
 import { Subject } from 'src/models/entities/subject.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
+import { SubjectEntity } from './entities/subject.entity';
 
 @Injectable()
 export class SubjectService {
@@ -11,28 +12,48 @@ export class SubjectService {
     @InjectRepository(Subject)
     private readonly subjectRepository: Repository<Subject>,
   ) {}
-  async create(createSubjectDto: CreateSubjectDto) {
+
+  async create(createSubjectDto: CreateSubjectDto): Promise<SubjectEntity> {
     try {
       const subject = this.subjectRepository.create(createSubjectDto);
-      return await this.subjectRepository.save(subject);
+      const savedSubject = await this.subjectRepository.save(subject);
+      return new SubjectEntity(savedSubject);
     } catch (error) {
       throw new InternalServerErrorException('An unexpected error occurred.');
     }
   }
 
-  async findAll() {
-    return await this.subjectRepository.find();
+  async findAll(): Promise<SubjectEntity[]> {
+    const subjects = await this.subjectRepository.find();
+    return subjects.map((subject) => new SubjectEntity(subject));
   }
 
-  async findOne(id: number) {
-    return await this.subjectRepository.findOneBy({ id });
+  async findOne(id: number): Promise<SubjectEntity> {
+    const subject = await this.subjectRepository.findOneBy({ id });
+    if (!subject) {
+      throw new InternalServerErrorException('Subject not found.');
+    }
+    return new SubjectEntity(subject);
   }
 
-  async update(id: number, updateSubjectDto: UpdateSubjectDto) {
-    return await this.subjectRepository.update(id, updateSubjectDto);
+  async update(
+    id: number,
+    updateSubjectDto: UpdateSubjectDto,
+  ): Promise<SubjectEntity> {
+    await this.subjectRepository.update(id, updateSubjectDto);
+    const updatedSubject = await this.subjectRepository.findOneBy({ id });
+    if (!updatedSubject) {
+      throw new InternalServerErrorException('Subject not found.');
+    }
+    return new SubjectEntity(updatedSubject);
   }
 
-  async remove(id: number) {
-    return await this.subjectRepository.delete(id);
+  async remove(id: number): Promise<SubjectEntity> {
+    const subject = await this.subjectRepository.findOneBy({ id });
+    if (!subject) {
+      throw new InternalServerErrorException('Subject not found.');
+    }
+    await this.subjectRepository.delete(id);
+    return new SubjectEntity(subject);
   }
 }
